Moved TypeIDs to a common header for both C & C++
