+++ symengine/basic.h
@@ -24,6 +24,7 @@
 #include <symengine/symengine_assert.h>
 #include <symengine/symengine_rcp.h>
 #include <symengine/dict.h>
+#include <symengine/type_codes.h>
 
 namespace SymEngine {
 
@@ -60,26 +61,6 @@ class Symbol;
     simpler code, you can use the add(), mul(), pow() functions that peform
     general and possibly slow canonicalization first.
 */
-
-enum TypeID {
-    INTEGER, RATIONAL, COMPLEX, COMPLEX_DOUBLE, REAL_MPFR, COMPLEX_MPC, REAL_DOUBLE,
-    // 'REAL_DOUBLE' returns the number of subclasses of Number.
-    // All subclasses of Number must be added before it. Do not assign
-    // non subclasses of Number before it.
-    SYMBOL, MUL, ADD, POW, UNIVARIATEPOLYNOMIAL, LOG, CONSTANT,
-    SIN, COS, TAN, COT, CSC, SEC,
-    ASIN, ACOS, ASEC, ACSC, ATAN, ACOT, ATAN2,
-    SINH, COSH, TANH, COTH,
-    ASINH, ACOSH, ATANH, ACOTH, ASECH,
-    LAMBERTW, ZETA, DIRICHLET_ETA, KRONECKERDELTA,
-    LEVICIVITA, GAMMA, LOWERGAMMA, UPPERGAMMA,
-    FUNCTIONSYMBOL, FUNCTIONWRAPPER, DERIVATIVE, SUBS, ABS,
-
-    // The 'TypeID_Count' returns the number of elements in 'TypeID'. For this
-    // to work, do not assign numbers to the elements above (or if you do, you
-    // must assign the correct count below).
-    TypeID_Count
-};
 
 class Basic {
 private:
@@ -292,4 +273,3 @@ const static TypeID type_code_id = ID; \
 virtual TypeID get_type_code() const { return type_code_id; };
 
 #endif
-
